♻️ Move E2E login test into user-account spec
Uses the existing fixture infrastructure (cookie banner, feature flags)
and the NEW_VISITOR_STATE override within the describe block.
